backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.4.2
certifi==2025.6.15
charset-normalizer==3.4.2
ckzg==1.0.2
click==8.2.1
cytoolz==1.0.1
dnspython==2.7.0
email_validator==2.2.0
eth-account==0.11.3
eth-hash==0.7.1
eth-keyfile==0.9.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.111.0
fastapi-cli==0.0.7
frozenlist==1.7.0
h11==0.16.0
hexbytes==0.3.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lru-dict==1.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
orjson==3.10.18
packaging==25.0
parsimonious==0.10.0
propcache==0.3.2
protobuf==6.31.1
py-ecc==8.0.0
py-solc-x==2.0.4
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
pyunormalize==16.0.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rlp==4.1.0
rpds-py==0.25.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
toolz==1.0.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
web3==6.19.0
websockets==15.0.1
yarl==1.20.1
